requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.3
pydantic-ai[openai]>=0.0.49
qdrant-client>=1.7.0
pydantic-settings==2.9.1
python-multipart
python-dotenv
pymysql==1.1.1
beautifulsoup4==4.12.3
openai>=1.0.0
jieba>=0.42.1
chonkie>=0.1.0
